[Feat] Add Collection Text
컬렉션에 마우스를 대면 툴팁이 나오는 기능 추가
