+++ Assets/PJW/Object/New Item.asset
@@ -10,14 +10,15 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 16df72ffc81be2f4f86baa453a28ce81, type: 3}
-  m_Name: HPPotion2
+  m_Name: New Item
   m_EditorClassIdentifier: 
-  ItemId: 999
-  Name: HPPotion2
-  Description: "HP \uD3EC\uC158\uC774\uB2E4"
+  ItemId: 11111
+  Name: 
+  Description: 
   Weight: 0
-  Type: 1
+  Type: 0
   MaxStackSize: 0
   Energy: 0
-  Icon: {fileID: 21300000, guid: b6b68c8c37d5cf1419267a80c426ac83, type: 3}
+  Icon: {fileID: 21300002, guid: 88dca061d041a2a4596563e6a38f66c8, type: 3}
   Prefab: {fileID: 0}
+  silhouetteSprite: {fileID: 21300000, guid: 87258a2e98839004986e33d9a04e825e, type: 3}
