+++ Assets/PJW/Scene/PJW_TestScene.unity
@@ -335,7 +335,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 4b72985614203aa42b10dc56d5f7fa86, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  itemData: {fileID: 11400000, guid: a48c438ee2e9a0446898d15ed199de9c, type: 2}
+  itemData: {fileID: 11400000, guid: 0a36c3d537d01ac4c98debf2a77b461a, type: 2}
 --- !u!114 &962111031
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -827,7 +827,7 @@ PrefabInstance:
     m_RemovedGameObjects: []
     m_AddedGameObjects:
     - targetCorrespondingSourceObject: {fileID: 3582733724437714125, guid: a2642c9ed05de3a4fac08e5f59e275b4, type: 3}
-      insertIndex: 1
+      insertIndex: 0
       addedObject: {fileID: 4675489283045035399}
     m_AddedComponents:
     - targetCorrespondingSourceObject: {fileID: 6850304826006328646, guid: a2642c9ed05de3a4fac08e5f59e275b4, type: 3}
@@ -1321,6 +1321,7 @@ MonoBehaviour:
   allItems:
   - {fileID: 11400000, guid: 431f32bb61874ce4180156e9dc49e6d9, type: 2}
   - {fileID: 11400000, guid: 37f1541bafa0e3b439bc4dd258998fc6, type: 2}
+  - {fileID: 11400000, guid: 0a36c3d537d01ac4c98debf2a77b461a, type: 2}
 --- !u!114 &4930864137705708120
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1462,7 +1463,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   gridRoot: {fileID: 2002097956294185741}
   entryUIPrefab: {fileID: 1499993659269505205, guid: a60fef34bbe63e04595a8574a0c0ad88, type: 3}
-  silhouetteSprite: {fileID: 21300000, guid: e55e4ef66043a8747866894761ac06be, type: 3}
 --- !u!1 &8342399958621166611
 GameObject:
   m_ObjectHideFlags: 0
