+++ Assets/PJW/Script/Collection/CollectibleItem.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class PJWCollectibleItem : MonoBehaviour
+public class CollectibleItem : MonoBehaviour
 {
     [SerializeField] private ItemSO itemData;
