--- Assets/PJW/Script/Collection/PJWItemEntryUI.cs
@@ -1,35 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-using UnityEngine.EventSystems;
-
-[RequireComponent(typeof(RectTransform))]
-public class PJWItemEntryUI : MonoBehaviour, IPointerEnterHandler, IPointerExitHandler
-{
-    [Header("UI References")]
-    [SerializeField] private Image icon;
-    [SerializeField] private GameObject tooltipPanel;
-    [SerializeField] private Text nameText;
-    [SerializeField] private Text descriptionText;
-
-    public void Initialize(ItemSO item, bool isCollected, Sprite silhouette)
-    {
-        nameText.text        = isCollected ? item.Name        : "???";
-        descriptionText.text = isCollected ? item.Description : "???";
-        icon.sprite          = isCollected ? item.Icon        : silhouette;
-        tooltipPanel.SetActive(false);
-    }
-
-    public void OnPointerEnter(PointerEventData eventData)
-    {
-        tooltipPanel.transform.SetAsLastSibling();
-        tooltipPanel.SetActive(true);
-    }
-
-    public void OnPointerExit(PointerEventData eventData)
-    {
-        tooltipPanel.SetActive(false);
-    }
-}
-
